requirements.txt
django == 3.2.8
djangorestframework